+++ test/test_runner.h
@@ -28,11 +28,17 @@
 
 /**
  * Get the parsing/API-success test array from unit-c.c.
+ *
+ * @param out_tests On success, set to pointer to the test array
+ * @param out_count On success, set to the number of tests
  */
 void get_unit_c_tests(const struct CMUnitTest** out_tests, size_t* out_count);
 
 /**
  * Get the NULL-dereference test array from unit-c-null.c.
+ *
+ * @param out_tests On success, set to pointer to the test array
+ * @param out_count On success, set to the number of tests
  */
 void get_unit_c_null_tests(const struct CMUnitTest** out_tests, size_t* out_count);